Fix: Improved Italian navigation prompts for street types
Adjusted the Italian translation of navigation instructions to use "in" instead of "sulla" when referring to streets. Previously, the navigator would incorrectly say "sulla viale" instead of "in viale". This change not only ensures grammatically correct translations but also provides a more natural and intuitive Italian phrasing, enhancing the user experience.

Signed-off-by: fparri <parrif_ibb@yahoo.com>